refactor: decompose smoke_utils.py into domain-specific sub-modules
Convert src/autoskillit/smoke_utils.py (1066-line monolith) into a smoke_utils/
package with 6 domain-specific sub-modules. An __init__.py facade re-exports all
19 public names, preserving the autoskillit.smoke_utils.function_name callable
path used by all recipe YAML files.

Sub-modules:
- _helpers.py: _load_json, try_load_json (shared JSON utilities)
- _review.py: annotate_pr_diff, check_review_loop, check_loop_iteration,
  check_loop_with_progress, enrich_diff_context, LOCAL_ROUND_EXEMPT_VERDICTS
- _eval.py: parse_eval_manifests, parse_agent_eval_manifests,
  build_eval_context, build_agent_eval_context, compile_eval_scorecard,
  _build_eval_context_common (consolidated helper replacing ~85% duplication)
- _telemetry.py: patch_pr_token_summary, _PR_SECTION_RE
- _git.py: check_bug_report_non_empty, compute_domain_partitions,
  fetch_merge_queue_data, detect_zero_changes, check_commits_ahead,
  close_issue_already_done

Dead temp_dir parameter removed from parse_eval_manifests.
