+++ src/autoskillit/smoke_utils/__init__.py
@@ -1,0 +1,52 @@
+"""Utility callables for smoke-test pipeline run_python steps.
+
+Known limitation: functions use hardcoded path conventions from the pipeline recipe.
+"""
+
+from autoskillit.smoke_utils._eval import (
+    build_agent_eval_context,
+    build_eval_context,
+    compile_eval_scorecard,
+    parse_agent_eval_manifests,
+    parse_eval_manifests,
+)
+from autoskillit.smoke_utils._git import (
+    check_bug_report_non_empty,
+    check_commits_ahead,
+    close_issue_already_done,
+    compute_domain_partitions,
+    detect_zero_changes,
+    fetch_merge_queue_data,
+)
+from autoskillit.smoke_utils._helpers import try_load_json
+from autoskillit.smoke_utils._review import (
+    LOCAL_ROUND_EXEMPT_VERDICTS,
+    annotate_pr_diff,
+    check_loop_iteration,
+    check_loop_with_progress,
+    check_review_loop,
+    enrich_diff_context,
+)
+from autoskillit.smoke_utils._telemetry import patch_pr_token_summary
+
+__all__ = [
+    "LOCAL_ROUND_EXEMPT_VERDICTS",
+    "annotate_pr_diff",
+    "build_agent_eval_context",
+    "build_eval_context",
+    "check_bug_report_non_empty",
+    "check_commits_ahead",
+    "check_loop_iteration",
+    "check_loop_with_progress",
+    "check_review_loop",
+    "close_issue_already_done",
+    "compile_eval_scorecard",
+    "compute_domain_partitions",
+    "detect_zero_changes",
+    "enrich_diff_context",
+    "fetch_merge_queue_data",
+    "parse_agent_eval_manifests",
+    "parse_eval_manifests",
+    "patch_pr_token_summary",
+    "try_load_json",
+]
